Set immutable rootfs and use dracut based standard initrd
This commit is two fold.

From one side it sets the use of dracut based
initrds and sets the cos-setup config to run after switching the root.
On the other hand this commit also sets the ephemeral /etc, /var and
/root mountpoints on tmpfs.

All cos-setup configurations are applied on
tmpfs after switching the root, right after local-fs.target (so all
the local filesystems are already there)

Fixes #40 related #5